Refactor : Posting 엔드포인트 변경 (postings -> posts) 및 줄바꿈 수정
